rename merge_candidate -> selectable_entity
