+++ app/modules/entities/components/selectable_entity.svelte
